cleaned up some bugs for semester tracker and dplist, as well as cleaned up some code in app.tsx
