+++ src/App.tsx
@@ -11,79 +11,9 @@ import { DpList } from "./DpList";
 import { AddDpModal } from "./AddDpModal";
 import { Course } from "./interfaces/course";
 import { Semester } from "./interfaces/semester";
+import { NewModal } from "./ NewModal";
 
 export function App(): JSX.Element {
-    const [semesters, setSemesters] = useState<Semester[]>([]);
-    const [selectedSemester, setSelectedSemester] = useState<string>("Fall");
-    const [newCourse, setNewCourse] = useState<Course>({
-        title: "",
-        courseCode: "",
-        credits: 0,
-        degreeRequirements: [],
-        coursePrereq: [],
-        courseCoreq: [],
-        courseDescription: ""
-    });
-
-    const addSemester = () => {
-        const newSemesterObj: Semester = {
-            id: semesters.length + 1,
-            title: selectedSemester,
-            courses: [],
-            totalCredits: 0
-        };
-        setSemesters([...semesters, newSemesterObj]);
-    };
-
-    const addCourse = (semesterId: number) => {
-        if (newCourse.courseCode && newCourse.title && newCourse.credits > 0) {
-            setSemesters((prevSemesters) =>
-                prevSemesters.map((semester) =>
-                    semester.id === semesterId
-                        ? {
-                              ...semester,
-                              // eslint-disable-next-line indent
-                              courses: [...semester.courses, newCourse]
-                          }
-                        : semester
-                )
-            );
-            setNewCourse({
-                courseCode: "",
-                title: "",
-                credits: 0,
-                degreeRequirements: [],
-                coursePrereq: [],
-                courseCoreq: [],
-                courseDescription: ""
-            });
-        }
-    };
-
-    const deleteSemester = (id: number) => {
-        const updatedSemesters = semesters.filter(
-            (semester) => semester.id !== id
-        );
-        setSemesters(updatedSemesters);
-    };
-
-    const deleteCourse = (semesterId: number, courseIndex: number) => {
-        setSemesters((prevSemesters) =>
-            prevSemesters.map((semester) =>
-                semester.id === semesterId
-                    ? {
-                          ...semester,
-                          courses: semester.courses.filter(
-                              (_, index) => index !== courseIndex
-                          )
-                      }
-                    : semester
-            )
-        );
-    };
-
-    const semesterOptions = ["Fall", "Winter", "Spring", "Summer"];
-
     //load in json data
     const DEGREEPLANS: DegreePlan[] = dpsamplejson.map(
         (dp: DegreePlan): DegreePlan => ({ ...dp }) //dp = degreeplan
@@ -158,127 +88,29 @@ export function App(): JSX.Element {
                     </p>
                 </header>
                 <DpList dp={degreePlans}></DpList>
-                <div />
-                <GenerateCSV
-                    data={[
-                        ["First Name", "Last Name"],
-                        ["Nicky", "Reigel"],
-                        ["Aidan", "Bell"]
-                    ]}
-                    filename="testexport"
-                />
-                <div />
-                <Import importData={importData} setImportData={setImportData} />
-                <div className="div-alightleft">
-                    <div className="App-blockleft">
-                        <p>sample block left</p>
-                    </div>
-                </div>
-                <div className="div-alignright">
-                    <div className="App-blockright">
-                        <p>sample block right</p>
-                    </div>
-                </div>
                 <Button className="add_btn" onClick={handleShowModal}>
                     Add New Degree Plan
                 </Button>
-                <AddDpModal
+                <Button onClick={saveData}>Save Degree Plans</Button>
+                <NewModal
                     show={showAddModal}
                     handleClose={handleCloseModal}
                     addDp={addDp}
-                ></AddDpModal>
-                <Button onClick={saveData}>Save Degree Plans</Button>
-            </div>
-            <div>
-                <h1>Semester Tracker</h1>
+                ></NewModal>
                 <div>
-                    <select
-                        value={selectedSemester}
-                        onChange={(e) => setSelectedSemester(e.target.value)}
-                    >
-                        {semesterOptions.map((option, index) => (
-                            <option key={index} value={option}>
-                                {option}
-                            </option>
-                        ))}
-                    </select>
-                    <button onClick={addSemester}>Add Semester</button>
+                    <Import
+                        importData={importData}
+                        setImportData={setImportData}
+                    />
+                    <GenerateCSV
+                        data={[
+                            ["First Name", "Last Name"],
+                            ["Nicky", "Reigel"],
+                            ["Aidan", "Bell"]
+                        ]}
+                        filename="testexport"
+                    />
                 </div>
-                <ul>
-                    {semesters.map((semester) => (
-                        <li key={semester.id}>
-                            <div>
-                                Semester: {semester.title}
-                                <button
-                                    onClick={() => deleteSemester(semester.id)}
-                                >
-                                    Delete Semester
-                                </button>
-                            </div>
-                            <div className="courses">
-                                {semester.courses.map((course, courseIndex) => (
-                                    <div key={courseIndex} className="course">
-                                        <span>
-                                            Course Code: {course.courseCode}
-                                            <br />
-                                            Course Title: {course.title}
-                                            <br />
-                                            Credits: {course.credits}
-                                        </span>
-                                        <button
-                                            onClick={() =>
-                                                deleteCourse(
-                                                    semester.id,
-                                                    courseIndex
-                                                )
-                                            }
-                                        >
-                                            Delete Course
-                                        </button>
-                                    </div>
-                                ))}
-                            </div>
-                            <div>
-                                <input
-                                    type="text"
-                                    value={newCourse.courseCode}
-                                    onChange={(e) =>
-                                        setNewCourse({
-                                            ...newCourse,
-                                            courseCode: e.target.value
-                                        })
-                                    }
-                                    placeholder="Course Code"
-                                />
-                                <input
-                                    type="text"
-                                    value={newCourse.title}
-                                    onChange={(e) =>
-                                        setNewCourse({
-                                            ...newCourse,
-                                            title: e.target.value
-                                        })
-                                    }
-                                    placeholder="Course Title"
-                                />
-                                <input
-                                    type="number"
-                                    value={newCourse.credits}
-                                    onChange={(e) =>
-                                        setNewCourse({
-                                            ...newCourse,
-                                            credits: +e.target.value
-                                        })
-                                    }
-                                    placeholder="Credits"
-                                />
-                                <button onClick={() => addCourse(semester.id)}>
-                                    Add Course
-                                </button>
-                            </div>
-                        </li>
-                    ))}
-                </ul>
             </div>
         </div>
     );
